test_start_stop_with_crash
